// dev-jobs-cron/detail.types.ts
// Generated by https://quicktype.io

import { Company, Img, CategoryTag } from "./job.types";

export interface JobDetail {
    job:                DescribeJob;
}

export interface DescribeJob {
    address:           Address;
    like_count:        number;
    id:                number;
    detail:            Detail;
    company_images:    CompanyImage[];
    skill_tags:        Tag[];
    status:            string;
    company:           Company;
    logo_img:          Img;
    company_tags:      Tag[];
    title_img:         Img;
    position:          string;
    category_tags:     CategoryTag[];
}

export interface Address {
    country:       string;
    full_location: string;
    geo_location:  GeoLocation;
    location:      string;
    country_code:  string;
}

export interface GeoLocation {
    n_location:    NLocation;
    location:      AddressDetail;
    location_type: string;
    bounds:        null;
}

export interface AddressDetail {
    lat: number;
    lng: number;
}

export interface NLocation {
    lat:        number;
    lng:        number;
    address:    string;
}

export interface CompanyImage {
    url: string;
    id:  number;
}

export interface Tag {
    title:      string;
    id:         number;
    kind_title: string;
}

export interface Detail {
    requirements:     string;
    main_tasks:       string;
    intro:            string;
    benefits:         string;
    preferred_points: string;
}